+++ typescript/infra/src/utils/safe.ts
@@ -14,7 +14,7 @@ import {
   getSafe,
   getSafeService,
 } from '@hyperlane-xyz/sdk';
-import { Address, CallData, eqAddress } from '@hyperlane-xyz/utils';
+import { Address, CallData, eqAddress, retryAsync } from '@hyperlane-xyz/utils';
 
 import safeSigners from '../../config/environments/mainnet3/safe/safeSigners.json' assert { type: 'json' };
 import { AnnotatedCallData } from '../govern/HyperlaneAppGovernor.js';
@@ -24,10 +24,10 @@ export async function getSafeAndService(
   multiProvider: MultiProvider,
   safeAddress: Address,
 ) {
-  const safeSdk: Safe.default = await getSafe(
-    chain,
-    multiProvider,
-    safeAddress,
+  const safeSdk: Safe.default = await retryAsync(
+    () => getSafe(chain, multiProvider, safeAddress),
+    5,
+    1000,
   );
   const safeService: SafeApiKit.default = getSafeService(chain, multiProvider);
   return { safeSdk, safeService };
@@ -39,6 +39,52 @@ export function createSafeTransactionData(call: CallData): MetaTransactionData {
     data: call.data.toString(),
     value: call.value?.toString() || '0',
   };
+}
+
+export async function executeTx(
+  chain: ChainNameOrId,
+  multiProvider: MultiProvider,
+  safeAddress: Address,
+  safeTxHash: string,
+): Promise<void> {
+  const { safeSdk, safeService } = await getSafeAndService(
+    chain,
+    multiProvider,
+    safeAddress,
+  );
+  const safeTransaction = await safeService.getTransaction(safeTxHash);
+  if (!safeTransaction) {
+    throw new Error(`Failed to fetch transaction details for ${safeTxHash}`);
+  }
+
+  // Throw if the safe doesn't have enough balance to cover the gas
+  let estimate;
+  try {
+    estimate = await safeService.estimateSafeTransaction(
+      safeAddress,
+      safeTransaction,
+    );
+  } catch (error) {
+    throw new Error(
+      `Failed to estimate gas for Safe transaction ${safeTxHash} on chain ${chain}: ${error}`,
+    );
+  }
+  const balance = await multiProvider
+    .getProvider(chain)
+    .getBalance(safeAddress);
+  if (balance.lt(estimate.safeTxGas)) {
+    throw new Error(
+      `Safe ${safeAddress} on ${chain} has insufficient balance (${balance.toString()}) for estimated gas (${
+        estimate.safeTxGas
+      })`,
+    );
+  }
+
+  await safeSdk.executeTransaction(safeTransaction);
+
+  console.log(
+    chalk.green.bold(`Executed transaction ${safeTxHash} on ${chain}`),
+  );
 }
 
 export async function createSafeTransaction(